Rename affiliates app to badges
